Delay activation of manual device select button
When the app first starts up there can be a slight delay before the
configuration files have been downloaded and adb starts up.

During this time, if the "Select device manually" button is pressed by
the user, the device selection combobox will be empty, which can be
confusing for the user.

This change disables the button until initialisation is complete,
displaying a "Loading device list..." message in the meantime.